rename&refactor: temperaly, a little err in ActivityChat.java
