+++ .idea/gradle.xml
@@ -12,7 +12,7 @@
           <set>
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/app" />
-            <option value="$PROJECT_DIR$/sendandreceive" />
+            <option value="$PROJECT_DIR$/chat_board" />
           </set>
         </option>
       </GradleProjectSettings>
