+++ chat_board/src/main/res/drawable/background_chat_input.xml
