+++ chat_board/src/main/res/drawable/background_content_top.xml
