+++ chat_board/src/main/res/drawable/background_input.xml
