+++ chat_board/src/main/res/drawable/background_received_message.xml
