+++ chat_board/src/main/res/drawable/background_sent_message.xml
