+++ chat_board/src/main/res/drawable/ic_send.xml
