+++ chat_board/src/main/res/layout/item_container_received_message.xml
