+++ chat_board/src/main/res/layout/item_container_sent_message.xml
