+++ chat_board/src/main/res/values/colors.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <color name="black">#FF000000</color>
+    <color name="white">#FFFFFFFF</color>
+    <color name="received_message_background">#090D16</color>
+    <color name="primary">#1C2E46</color>
+    <color name="primary_dark">#142232</color>
+    <color name="Primary_text">#212121</color>
+    <color name="secondary_text">#757575</color>
+    <color name="error">#B00020</color>
+    <color name="input_background">#ECECEC</color>
+</resources>
